// src/api/logistic/config-salesorg-customer-group-material/types.ts
export interface ListSalesORGCustomerGroupMaterial {
    readonly company_id: string
    readonly sales_org_id: string
    readonly sales_org_name: string
    readonly customer_group_id: string
    readonly customer_group_name: string
    readonly product_id: string
    readonly product_name: string
    readonly min_qty: string
    readonly max_qty: string
    readonly uom_id: string
    readonly status: string
  }